Further development on contract refactor
Here I'm introducing the escrow contract. The idea here is that we need a separate contract for each purchase to hold funds. Technically we don't have to deploy a new contract, but this is a security mechanism so that we don't have one contract that could potentially be holding funds for many transactions at once.