NewEntry events include the entryType id.
If you are following `NewEntry` events, you are doing so in order to look up details about those entries. If you have an entryType id along with your entry id, then you can look up the contract address in a local cache, and immediately ask the correct contract for details about the new entry.

Otherwise, you have to make an extra call to the the registry for every entry to find out which contract it lives on.