Add pause & unpause to token_cli
This adds the `pause` and `unpause` actions to the token CLI, which
pause and unpause token transfers. These may only be called by the token
contract owner.

This also adds `sendTransaction` to the token library, which sends an
Ethereum transaction and polls for the receipt. This is necessary
workaround for an issue where the normal `contract.methods.MyMethod.send()`
function never returns a transaction receipt for certain calls:

https://github.com/INFURA/infura/issues/95

Manually tested the token faucet on Rinkeby & localhost to ensure that it still
works.